Factor out ExtensibleLoadManagerImplTest setup to a base class
